Move derivation of evspsblpot to separate file
