+++ esmvaltool/diag_scripts/hydrology/__init__.py
@@ -1,0 +1,1 @@
+"""Collection of scripts to prepare hydrological model forcings."""
